Control.LVish.CancelT

Description

A module for adding the cancellation capability.

In its raw form, this is unsafe, because cancelating could cancel something that would have performed a visible side effect.

The transformer that adds the cancellation capability

data CancelT m a

A Par-monad scheduler transformer that adds the cancellation capability. To do this, it must track, and periodically poll, extra mutable state for each cancellable computation in the fork-tree.

type ThreadId = CState

The type of cancellable thread identifiers.

data CFut m a

Futures that may be canceled before the result is available.

data CFutFate

In a deterministic scenario, a CFut can only be canceled or read, not both.

Operations specific to CancelT

runCancelT :: ParMonad m => CancelT m a -> m a

Run a Par monad with the cancellation effect. Within this computation, it is possible to cancel subtrees of computations.

Forking computations that may be canceled

forkCancelable :: (ParIVar m, LVarSched m, ReadOnlyM m, FutContents m CFutFate, FutContents m a) => CancelT m a -> CancelT m (ThreadId, CFut m a)

Fork a computation while retaining a handle on it that can be used to cancel it (and all its descendents). This is equivalent to createTid followed by forkCancelableWithTid.

This version is expected to retain determinism. Therefore, the canceled computations must be read only.

Finally, note that the return value

forkCancelableND :: (ParIVar m, LVarSched m, HasIOM m, FutContents m CFutFate, FutContents m a) => CancelT m a -> CancelT m (ThreadId, CFut m a)

This is a version of forkCancelable which allows side-effecting computation to be canceled. Be warned, that is a dangerous business. Accordingly, this function introduces nondeterminism and must register the IO.

More detailed interface, separating forking and TID creation

createTid :: (ParMonad m, LVarSched m) => CancelT m ThreadId

Sometimes it is necessary to have a TID in scope *before* forking the computations in question. For that purpose, we provide a two-phase interface: createTid followed by forkCancelableWithTid.

forkCancelableWithTid :: (ParIVar m, LVarSched m, ReadOnlyM m, FutContents m CFutFate, FutContents m a) => ThreadId -> CancelT m a -> CancelT m (CFut m a)

Add a new computation to an existing cancelable thread.

Forking multiple computations with the same Tid are permitted; all threads will be canceled as a group.

forkCancelableNDWithTid :: (ParIVar m, LVarSched m, HasIOM m, FutContents m CFutFate, FutContents m a) => ThreadId -> CancelT m a -> CancelT m (CFut m a)

Variant of forkCancelableWithTid that works for nondeterministic computations.

Performing cancellation, or waiting for results

readCFut :: (ParIVar m, HasPut (GetEffects m), FutContents m CFutFate, FutContents m a) => CFut m a -> m a

Do a blocking read on the result of a cancelable-future. This is an ERROR if the future has been cancelled. (And for determinism, the reverse is also true. A cancellation after this read is an error.)

cancel :: (ParMonad m, Monad m, HasPut (GetEffects m)) => ThreadId -> CancelT m ()

Issue a cancellation request for a given sub-computation. It will not be fulfilled immediately, because the cancellation process is cooperative and only happens when the thread(s) check in with the scheduler.

pollForCancel :: (ParMonad m, LVarSched m) => CancelT m ()

Check with the scheduler to see if the current thread has been canceled, if so stop computing immediately.

cancelMe :: LVarSched m => CancelT m ()

Self cancellation. Not the same as the parent calling cancel, because cancelMe only cancels the current leaf of the computation tree, not its siblings (not the entire subtree under a forkCancelable, that is).

Boolean operations with cancellation

asyncAnd :: forall p. (ParIVar p, LVarSched p, ReadOnlyM p, FutContents p CFutFate, FutContents p ()) => CancelT p Bool -> CancelT p Bool -> (Bool -> CancelT p ()) -> CancelT p ()

A parallel and operation that not only signals its output early when it receives a False input, but also attempts to cancel the other, unneeded computation.

FIXME: This is unsafe until there is a way to guarantee that read-only computations are performed!