Control.Par.ST

Description

This file provides a basic capability for parallel in-place modification of (disjoint) partitions of an array. It allows a Par computation to carry an implicit vector in the background, which allows mutation via arbitrary ST computations.

This module does NOT provide a monad-transformer. Rather, it is a ROOT for a monad-transformer stack, and is hard-wired to use the Control.LVish version of the Par monad underneath.

The monad: a dischargable effect

data ParST stState parM ans

The ParST monad. It uses the StateT monad transformer to layer a state of type on top of an inner monad, Par.

The first type parameter determines the type of state held. The parameters det and s2 are for the underlying Par monad.

Its final parameter, ans, is the result of running the entire computation, after which the vector is no longer accessible.

runParST :: forall stt s0 parM ans. ParThreadSafe parM => stt s0 -> (forall s1. ParST (stt s1) parM ans) -> parM ans

runParST discharges the extra state effect leaving the the underlying Par computation only -- just like runStateT. Here, using the standard trick runParST has a rank-2 type, with a phantom type s1.

An alternate fork operation

forkSTSplit

Arguments

:: forall a b s0 parM stt . (ParThreadSafe parM, ParFuture parM, FutContents parM a, Eq a, STSplittable stt)

=> SplitIdx stt
Where to split the data.

-> (forall sl. ParST (stt sl) parM a)
Left child computation.

-> (forall sr. ParST (stt sr) parM b)
Right child computation.

-> ParST (stt s0) parM (a, b)

forkWithVec takes a split point and two ParST computations. It gets the state of the current computation, for example a vector, and then divides up that state between the two other computations.

Writes in those two computations may actually mutate the original vector. But forkWithVec is a fork-join construct, rather than a one-sided fork such as fork. So the continuation of forkWithVec will not run until both child computations return, and are thus done accessing the state.

Working with ST and other lifts

liftST :: ParThreadSafe parM => ST s1 a -> ParST (stt s1) parM a

Allow ST computations inside ParST computations. This operation has some overhead.

liftPar :: ParThreadSafe parM => parM a -> ParST stt1 parM a

Lift an ordinary Par computation into ParST.

Convert between state types

transmute :: forall a b s parM ans. (ParThreadSafe parM, STSplittable a, STSplittable b) => (b s -> a s) -> ParST (a s) parM ans -> ParST (b s) parM ans

Type class for valid states.

class STSplittable ty where

The class of types that can be modified in ST computations, and whose state can be partitioned into disjoint pieces to be passed linearly to exactly one parallel subcomputation.

Associated Types

type SplitIdx ty :: *

Something of type SplitIdx describes where and how to split the data into two pieces.

Methods

splitST :: SplitIdx ty -> ty s -> (ty s, ty s)

splitST does the actual splitting.

Annoying newtypes and wrappers to take the s param last:

newtype MVectorFlp a s

The ways to split a vector. For now we only allow splitting into two pieces at a given index. In the future, other ways of partitioning the set of elements may be possible. newtype VecSplit = SplitAt Int

An annoying type alias simply for the purpose of arranging for the s parameter to be last.

Constructors

VFlp (MVector s a)

data UVectorFlp a s

An annoying type alias simply for the purpose of arranging for the s parameter to be last. Also carries the Unbox constraint.

Constructors

Unbox a => UFlp (MVector s a)

data SVectorFlp a s

An annoying type alias simply for the purpose of arranging for the s parameter to be last. Also carries the Storable constraint.

Constructors

Storable a => SFlp (MVector s a)

data STTup2 a b s

An annoying type wrapper simply for the purpose of arranging for the s parameter to be last.

Constructors

STTup2 !(a s) !(b s)

data STUnit s

A splittable type which contains no information.

Constructors

STUnit