Control.Par.StateT

Description

This module provides a notion of (Splittable) State that is compatible with any Par monad.

This module provides instances that make StateT-transformed monads into valid Par monads.

Class for state split at fork points

class SplittableState a where

A type in SplittableState is meant to be added to a Par monad using StateT. It works like any other state except at fork points, where the runtime system splits the state using splitState.

Common examples for applications of SplittableState would include (1) routing a splittable random number generator through a parallel computation, and (2) keeping a tree-index that locates the current computation within the binary tree of forks. Also, it is possible to simply duplicate the state at all fork points, enabling thread local copies of the state.

The limitation of this approach is that the splitting method is fixed, and the same at all fork points.

Methods

splitState :: a -> (a, a)

Also note exported instances.